+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v7.0.1-1.20] - 2023-06-13
+## [v7.0.1-1.20] - 2023-06-17
 ### Changed
 - Don't require enum values field to be final, some mixin accessor might have removed that flag
+### Fixed
+- Fixed `RaiderType` factory having incorrect entity type bounds
+- Fixed `IllagerSpell` factory not updating an internal enum values by id function
 
 ## [v7.0.0-1.20] - 2023-06-10
 - Ported to Minecraft 1.20